Bug #2030493 64bit bsp sprd_cpp_addr
[root cause  ]: bsp sprd_cpp_addr
[changes     ]: bsp sprd_cpp_addr
[side effects]:no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:n6pro
[test case]:camera
[reviewers]:hongcheng.li

commit_template_version:v1
[change_type ] feature
[tag_product ] common

+++ interface/sprd_cpp.h
@@ -269,10 +269,10 @@ struct sprd_cpp_trim {
 };
 
 struct sprd_cpp_addr {
-	unsigned int y;
-	unsigned int u;
-	unsigned int v;
-	unsigned int mfd[3];
+	unsigned long long y;
+	unsigned long long u;
+	unsigned long long v;
+	unsigned long long mfd[3];
 };
 
 struct sprd_cpp_rot_cfg_parm {
